LayoutTests/fast/dom/Node/textContent-mutationEvents-expected.txt

Tests that the DOM mutation events are fired correctly when Node.textContent is set.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


title.textContent = 'New title';

* DOMNodeRemoved fired
PASS firedDOMNodeRemoved is false
PASS firedDOMNodeInserted is false
PASS firedDOMSubtreeModified is false
PASS document.title is "Original title"

* DOMNodeInserted fired
PASS firedDOMNodeRemoved is true
PASS firedDOMNodeInserted is false
PASS firedDOMSubtreeModified is false
PASS document.title is "New title"

* DOMSubtreeModified fired
PASS firedDOMNodeRemoved is true
PASS firedDOMNodeInserted is true
PASS firedDOMSubtreeModified is false
PASS document.title is "New title"

PASS firedDOMNodeRemoved is true
PASS firedDOMNodeInserted is true
PASS firedDOMSubtreeModified is true
PASS document.title is "New title"
PASS firedMutationObserver is false

* Mutation observer invoked
PASS firedMutationObserver is false
PASS testMutations.length is 1
PASS mutation.type is "childList"
PASS mutation.target is title
PASS mutation.removedNodes.length is 1
PASS mutation.removedNodes[0].data is "Original title"
PASS mutation.addedNodes.length is 1
PASS mutation.addedNodes[0].data is "New title"
PASS successfullyParsed is true

TEST COMPLETE
